python_ai/requirements.txt
# Core ML/AI
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
google-generativeai>=0.3.0

# Resume Parsing
pdfminer.six>=20221105
python-docx>=0.8.11

# API
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Optional: For better NLP (install separately)
# pip install spacy sentence-transformers torch